These forms are going to be the end of me!
I've added aria-checked attributes to the div wrapped around each input and to their labels so that no matter what the user clicks on the "selected" or "unselected" value should be read to them.

Also, the description text should only appear in one place (not as the title of the tooltip, only as text) and not with aria-hidden which keeps it from being read.

I think because the tooltip was described by an aria-hidden span that the tooltip wasn't always being read-aloud by screen readers.